+++ natural_selection/genetic_programs/__init__.py
@@ -28,7 +28,7 @@ def random_generate(operators : list = None,
                     max_depth : int = None,
                     min_depth : int = None,
                     growth_mode : str = 'grow',
-                    terminal_prob : float = 0.5,
+                    terminal_prob : float = None,
                     genetic_program = None):
     """
     Function for random node tree creation. This function can be used for either Full tree generation or Grow generation.
@@ -62,7 +62,11 @@ def random_generate(operators : list = None,
             raise GeneticProgramError('No max_depth or genetic program given')
 
     def create_random_node_or_terminal():
-        if random.random() > terminal_prob:
+        if terminal_prob:
+            prob = terminal_prob
+        else:
+            prob = genetic_program.terminal_prob
+        if random.random() > prob:
             return create_random_node()
         return create_random_terminal()
 
@@ -170,8 +174,10 @@ class Node:
         if children:
             self.children = children
             self.arity = len(children)
-        else:
+        elif not is_terminal:
             self.children = [None] * self.arity
+        else:
+            self.children = list()
 
     def __call__(self, **kwargs):
         if self.is_terminal:
@@ -270,6 +276,54 @@ class Node:
                 deepest = d
         return deepest
 
+    def __breadth(self, to_depth : int , current_depth : int = 0):
+
+        if current_depth == to_depth-1:
+            return len(self.children)
+        else:
+            b = 0
+            for n in self.children:
+                b += n.__breadth(to_depth=to_depth, current_depth=current_depth+1)
+            return b
+
+    def breadth(self, depth: int = None):
+        """
+        Find the breadth at a given depth of the tree. If no depth is given (None), default to the trees depth.
+
+        Args:
+            depth (int): At which depth to find the breadth (default = None).
+
+        Returns:
+            int: The breadth of the given depth.
+        """
+        if depth is None:
+            depth = self.depth()
+        if depth > self.depth() or depth < 1:
+            raise GeneticProgramError(f'Current tree depth not reachable at {depth}')
+        return self.__breadth(to_depth=depth, current_depth=1)
+
+    def max_breadth(self):
+        """
+        Find the most broadest part of the tree, returning the breadth and depth at which it occurs.
+
+        Returns:
+            tuple: Breadth, depth
+        """
+        depth = self.depth()
+        max_breadth = 0
+        max_breadth_depth = 0
+
+        for i in range(depth):
+            b_at = self.breadth(i+1)
+            if b_at > max_breadth:
+                max_breadth_depth = i+1
+                max_breadth = b_at
+
+        return max_breadth, max_breadth_depth
+
+
+
+
 class GeneticProgram:
     """
     A class that encapsulates a single genetic program, with node tree and a fitness evaluation function.
@@ -336,6 +390,7 @@ class GeneticProgram:
             terminal_prob=terminal_prob,
             genetic_program=self
         )
+        self.terminal_prob = terminal_prob
         self.growth_mode = growth_mode
         self.tree_generator = tree_generator
         self.root_node = Node(label=self.name,arity=1, operator=op.OperatorReturn())
@@ -363,6 +418,14 @@ class GeneticProgram:
             raise GeneticProgramError('No node tree has been constructed!')
         return self.root_node(**kwargs)
 
+    def depth(self):
+        """
+        Wrapping around the depth function of the node tree.
+
+        Returns:
+            int: Node tree depth.
+        """
+        return self.node_tree.depth()
 
     def register_parent_names(self, parents: list, reset_parent_name_list: bool = True):
         """