+++ docs/source/changelog.rst
@@ -17,6 +17,7 @@ Version 0.2.26 (2022-)
 * Added tree generation function ``random_generate`` to genetic_programs.
 * Renamed ``natural_selection.genetic_programs.functions `` to ``natural_selection.genetic_programs.node_operators ``.
 * Added ``is_empty`` to Node class.
+* Added ``breadth`` and ``max_breadth`` to Node class, useful for doing crossover.
 * Added more history "stamps" to Island for alien spawn, migrant import.
 * General work on ``GeneticProgram`` class.
 * Added ``create_genetic_program`` to Island for easy wrapper.